+++ packages/platform-sdk-profiles/__tests__/env.test.ts
@@ -1,7 +1,7 @@
 import "jest-extended";
 
 import { Environment } from "../src/env";
-import { Profiles } from "../src/profiles";
+import { ProfileRepository } from "../src/profile-repository";
 import { HttpClient } from "./stubs/client";
 
 let subject: Environment;
@@ -9,5 +9,5 @@ let subject: Environment;
 beforeEach(() => (subject = new Environment({ httpClient: new HttpClient(), storage: "localstorage" })));
 
 test("Environment#profiles", async () => {
-	expect(subject.profiles()).toBeInstanceOf(Profiles);
+	expect(subject.profiles()).toBeInstanceOf(ProfileRepository);
 });