+++ packages/platform-sdk-profiles/src/profile-repository.ts
@@ -4,7 +4,7 @@ import { v4 as uuidv4 } from "uuid";
 import { Storage } from "./contracts";
 import { Profile } from "./profile";
 
-export class Profiles {
+export class ProfileRepository {
 	readonly #key: string = "profiles";
 
 	readonly #httpClient: Contracts.HttpClient;
@@ -57,6 +57,7 @@ export class Profiles {
 			id: uuidv4(),
 			name,
 			wallets: [],
+			contacts: [],
 			httpClient: this.#httpClient,
 			storage: this.#storage,
 		});